Draft update to our builder

+++ lang/shell/bash/functions/common.sh
@@ -3113,12 +3113,15 @@ koopa_brew_upgrade_brews() {
 }
 
 koopa_build_all_apps() {
-    local app pkg pkgs
+    local app dict pkg pkgs
     koopa_assert_has_no_args "$#"
     declare -A app=(
         [koopa]="$(koopa_locate_koopa)"
     )
     [[ -x "${app[koopa]}" ]] || return 1
+    declare -A dict=(
+        [opt_prefix]="$(koopa_opt_prefix)"
+    )
     pkgs=()
     pkgs+=('pkg-config' 'make')
     koopa_is_linux && pkgs+=('attr')
@@ -3203,6 +3206,7 @@ koopa_build_all_apps() {
         'proj'
         'gdal'
         'fribidi'
+        'harfbuzz'
         'r'
         'conda'
         'apr'
@@ -3247,8 +3251,6 @@ koopa_build_all_apps() {
         'groff'
         'gsl'
         'gzip'
-        'harfbuzz'
-        'hyperfine'
         'oniguruma'
         'jq'
         'less'
@@ -3283,7 +3285,6 @@ koopa_build_all_apps() {
         'which'
         'libgeotiff'
         'go'
-        'apptainer'
         'chezmoi'
         'fzf'
         'aws-cli'
@@ -3321,6 +3322,7 @@ koopa_build_all_apps() {
         'difftastic' # deps: rust
         'du-dust' # deps: rust
         'exa' # deps: rust
+        'hyperfine' # deps: rust
         'mcfly' # deps: rust
         'mdcat' # deps: rust
         'procs' # deps: rust
@@ -3341,7 +3343,6 @@ koopa_build_all_apps() {
     then
         pkgs+=(
             'anaconda'
-            'aspera-connect'
             'docker-credential-pass'
             'hadolint'
             'haskell-stack'
@@ -3351,9 +3352,26 @@ koopa_build_all_apps() {
             'snakemake'
         )
     fi
-    koopa_is_linux && pkgs+=('lmod')
+    if koopa_is_linux
+    then
+        pkgs+=(
+            'apptainer'
+            'lmod'
+        )
+        if ! koopa_is_aarch64
+        then
+            pkgs+=(
+                'aspera-connect'
+            )
+        fi
+    fi
     for pkg in "${pkgs[@]}"
     do
+        if [[ -L "${dict[opt_prefix]}/${pkg}" ]] && \
+            [[ -e "${dict[opt_prefix]}/${pkg}" ]]
+        then
+            continue
+        fi
         "${app[koopa]}" install "$pkg"
     done
     koopa_push_all_apps
